requirements.txt
# Core agent dependencies
langchain==1.2.7
langchain-google-genai==4.2.0
langgraph==1.0.7
langchain-mcp-adapters==0.1.11
python-dotenv
neo4j

# Model Judge
groq
